Split exceptions into separate files
