+++ harmony-kotlin/src/commonMain/kotlin/com.harmony.kotlin/error/UnauthorizedException.kt
@@ -1,0 +1,6 @@
+package com.harmony.kotlin.error
+
+/**
+ * Not authorized to access the data
+ */
+class UnauthorizedException(message: String? = null, cause: Throwable, isResolved: Boolean) : HarmonyException(message, cause)
